refactor: layout view
